Move new conversation sticky to top
