+++ web/screens/Chat/index.tsx
@@ -10,10 +10,8 @@ const ChatScreen = () => {
   return (
     <div className="flex h-full">
       <div className="flex h-full w-64 flex-shrink-0 flex-col overflow-y-auto border-r border-border ">
-        <div className="px-4 py-6 pt-4">
-          <LeftHeaderAction />
-          <HistoryList />
-        </div>
+        <LeftHeaderAction />
+        <HistoryList />
       </div>
       <div className="relative flex h-full w-full flex-col bg-background/50">
         <MainHeader />
